refactored opening map logic into it's own file / hook
