+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/location/location.tsx
@@ -136,7 +136,10 @@ const LocationInput = ({ onChange, value, errorListener }: any) => {
   const { listenTo, stopListening } = useBackbone()
   const { MuiButtonProps, MuiPopoverProps } = useMenuState()
   const onDraw = () => {
-    ;(wreqr as any).vent.trigger('search:draw' + state.mode, locationModel)
+    ;(wreqr as any).vent.trigger(
+      'search:draw' + locationModel.attributes.mode,
+      locationModel
+    )
   }
   const onDrawCancel = () => {
     ;(wreqr as any).vent.trigger('search:drawcancel', locationModel)
